backend/requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
pandas
openpyxl
scikit-learn
joblib
numpy
bcrypt
python-dotenv
django-cors-headers
